Catch debconf error
Using the Debian image, you always end up with the following error
printed to stderr:

    debconf: delaying package configuration, since apt-utils is not
    installed

It seems like this is a problem in Debian which causes no actual
problems but is also unlikely to be fixed soon. That is why this patch
just catches the problems and turns this into an info level log entry
instead.